requirements.txt
# baidu-aip==2.2.18.0
certifi==2020.12.5
chardet==3.0.4
idna==2.7
Pillow==8.3.2
PySocks==1.7.1
requests==2.20.0
urllib3==1.26.5
xlrd==1.2.0
